Remove onAction from feedback
